upath._protocol: move to upath protocol parsing
